// src/interfaces/IAccountTranslate.ts
export interface IAccountTranslate {
  personalInfo: string;
  fullname: string;
  phone: string;
  none: string;
  gender: string;
  birthDay: string;
  yearOfBirth: string;
  address: string;
  update: string;
  changePassword: string;
  oldPassword: string;
  newPassword: string;
  rePassword: string;
  edit: string;
  deleteAccount: string;
  editInformation: string;
  save: string;
  saveInfomation: string;
  back: string;
  updateSuccess: string;
  male: string;
  female: string;
  other: string;
  deleteContent: string;
  confirm: string;
}